pyproject.toml
[tool.poetry]
name = "internetum "
version = "0.1.0"
description = ""
authors = ["Fredrik Svärd <nfredrik@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
behave = "^1.2.6"
requests = "^2.31.0"
ping3 = "^4.0.4"
scapy = "^2.5.0"
assertpy = "^1.1"
netifaces = "^0.11.0"
python-dateutil = "^2.8.2"
pydantic = "^2.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

